backend/infra/pyproject.toml
[project]
name = "splink-infra"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.0.0",
]
